+++ src/main/kotlin/com/github/rebel000/cmdlineargs/Globals.kt
@@ -1,0 +1,12 @@
+package com.github.rebel000.cmdlineargs
+
+import org.jetbrains.annotations.NonNls
+
+@NonNls
+const val COMMANDLINE_PATCHED = "com.github.rebel000.cmdlineargs.patched"
+
+@NonNls
+const val TOOLWINDOW_ID = "cmdlineargs"
+
+@NonNls
+const val TOOLBAR_ACTION_ID = "cmdlineargs.actions"