+++ src/main/kotlin/com/github/rebel000/cmdlineargs/treeactions/CopyAction.kt
@@ -1,24 +1,15 @@
 package com.github.rebel000.cmdlineargs.treeactions
 
-import com.github.rebel000.cmdlineargs.Resources
 import com.github.rebel000.cmdlineargs.ui.ArgumentTree
-import com.intellij.icons.AllIcons
-import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.IdeActions
 import com.intellij.openapi.ide.CopyPasteManager
 import java.awt.datatransfer.StringSelection
 import java.util.*
 
-@Suppress("DialogTitleCapitalization")
-class CopyAction(private val tree: ArgumentTree) :
-    AnAction(Resources.message("action.copy"), Resources.message("action.copy"), AllIcons.Actions.Copy) {
-    init {
-        registerCustomShortcutSet(ActionManager.getInstance().getAction(IdeActions.ACTION_COPY).shortcutSet, tree)
-    }
-
+class CopyAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
+        val tree = ArgumentTree.getInstance(e.project) ?: return
         val nodes = tree.selectedNodes(false)
         val strings = Vector<String>()
         for (node in nodes) {