+++ src/main/kotlin/com/github/rebel000/cmdlineargs/treeactions/EditAction.kt
@@ -1,11 +1,12 @@
 package com.github.rebel000.cmdlineargs.treeactions
 
 import com.github.rebel000.cmdlineargs.ui.ArgumentTree
-import com.intellij.ui.AnActionButton
-import com.intellij.ui.AnActionButtonRunnable
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
 
-class EditAction(private val tree: ArgumentTree) : AnActionButtonRunnable {
-    override fun run(button: AnActionButton) {
+class EditAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        val tree = ArgumentTree.getInstance(e.project) ?: return
         val node = tree.selectedNode()
         if (node != null) {
             tree.editNode(node)