+++ src/main/kotlin/com/github/rebel000/cmdlineargs/treeactions/MoveUpAction.kt
@@ -1,14 +1,13 @@
 package com.github.rebel000.cmdlineargs.treeactions
 
 import com.github.rebel000.cmdlineargs.ui.ArgumentTree
-import com.github.rebel000.cmdlineargs.ui.ArgumentTreeNode
-import com.intellij.ui.AnActionButton
-import com.intellij.ui.AnActionButtonRunnable
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
 import javax.swing.tree.TreePath
-import kotlin.math.min
 
-class MoveUpAction(private val tree: ArgumentTree) : AnActionButtonRunnable {
-    override fun run(button: AnActionButton) {
+class MoveUpAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        val tree = ArgumentTree.getInstance(e.project) ?: return
         val selectedNodes = tree.selectedNodes(true)
         if (selectedNodes.isNotEmpty()) {
             val newSelectionPaths = ArrayList<TreePath>(selectedNodes.count())
@@ -22,12 +21,10 @@ class MoveUpAction(private val tree: ArgumentTree) : AnActionButtonRunnable {
                     if (neighbor?.isFolder == true) {
                         parent = neighbor
                         index = neighbor.childCount
-                    }
-                    else {
+                    } else {
                         index--
                     }
-                }
-                else {
+                } else {
                     val grandparent = parent.parent ?: return
                     index = grandparent.getIndex(parent)
                     parent = grandparent
@@ -36,6 +33,9 @@ class MoveUpAction(private val tree: ArgumentTree) : AnActionButtonRunnable {
                     val wasExpanded = tree.isExpanded(TreePath(node.path))
                     tree.removeNode(node)
                     tree.insertNode(node, parent, index)
+                    if (parent.singleChoice) {
+                        tree.setNodeState(node, false)
+                    }
                     val path = TreePath(node.path)
                     newSelectionPaths.add(path)
                     if (wasExpanded) {