+++ src/main/kotlin/com/github/rebel000/cmdlineargs/treeactions/ReloadAction.kt
@@ -1,14 +1,12 @@
 package com.github.rebel000.cmdlineargs.treeactions
 
-import com.github.rebel000.cmdlineargs.Resources
 import com.github.rebel000.cmdlineargs.ui.ArgumentTree
-import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 
-class ReloadAction(private val tree: ArgumentTree) :
-    AnAction(Resources.message("action.reload"), Resources.message("action.reload"), AllIcons.Actions.Refresh) {
+class ReloadAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
+        val tree = ArgumentTree.getInstance(e.project) ?: return
         tree.reloadState()
     }
 }